add sample code for get spring bean in other context Servlet or Filter

+++ src/main/java/com/sample/repo/dao/PersonDAO.java
@@ -9,6 +9,6 @@ public interface PersonDAO {
 	public void addPerson(Person p);
 	public void updatePerson(Person p);
 	public List<Person> listPersons();
-	public Person getPersonById(int id);
-	public void removePerson(int id);
+	public Person getPersonById(long id);
+	public void removePerson(long id);
 }
