+++ src/main/java/com/sample/repo/service/PersonService.java
@@ -9,7 +9,7 @@ public interface PersonService {
 	void addPerson(Person p);
 	void updatePerson(Person p);
 	List<Person> listPersons();
-	Person getPersonById(int id);
-	void removePerson(int id);
+	Person getPersonById(long id);
+	void removePerson(long id);
 	
 }
